Used typed Expression.observedValue extensions commit_hash:397afc589d4b61774777c9bd78c188b66ea83ac6

+++ client/android/compose/src/main/kotlin/com/yandex/div/compose/views/modifiers/SizeModifiers.kt
@@ -88,6 +88,5 @@ private fun DivSize.WrapContent.wrapContentSizeBounds(): Pair<Dp?, Dp?> {
 
 @Composable
 private fun DivSizeUnitValue.toDpSize(): Dp {
-    return value.observedValue().toFloat().toDp(unit.observedValue())
+    return value.observedValue().toDp(unit.observedValue())
 }
-
